Shift CreateOrReadOutputFile method from Class Task to Class Storage
