+++ src/main/java/omoh/Storage.java
@@ -1,6 +1,11 @@
 package omoh;
 
+import omoh.customexceptions.CorruptedFileException;
+import omoh.customexceptions.EmptyTodoException;
+import omoh.tasktypes.Task;
+
 import java.io.File;
+import java.io.FileNotFoundException;
 
 public class Storage {
 
@@ -10,6 +15,22 @@ public class Storage {
         File directory = new File("data");
         if (!directory.exists()) {
             directory.mkdirs(); // This will create the necessary directories
+        }
+    }
+
+    //creates output.txt file to store data
+    //also reads output file
+    public static void createOrReadOutputFile () {
+        //code to create new file in data directory
+        File f = new File("data/output.txt");
+        try {
+            Task.readFile();
+        } catch (FileNotFoundException e) {
+            System.out.println("File not found");
+        } catch (EmptyTodoException e) {
+            System.out.println("Todo task empty");
+        } catch (CorruptedFileException e) {
+            System.out.println("Output.txt file format is wrong. File corrupted.");
         }
     }
 }
